Explicit C++-side client state
This adds a client state definition to the C++ code, including some rudimentary
management logic. The state is attached to the Client object and extensible in
Lua, i.e. it can be used as a normal table for fields that are not pre-defined.

The client functionality has been changed from RAII to explicit initialization
via connect() and close(). This makes life easier in Lua since the same client
and state object can be passed around and will remaining valid after
re-connects.