Active unit tracking and end-of-battle detection in client::State
This ports the remaining logic from torchcraft:receive() to the C++ State
object. In particular, client::State now keeps track of units that are alive,
similar to what the Lua code has been doing previously with `units_myself`,
`units_enemy` and so on. Unfortunately, this requires maintaining a
near-duplicate of frame->units so that client code is able to work with the
correct units if frame skipping is enabled on the server (and thus frame->units
might include units that have died since the previous update).

Another significant change is that the `torchcraft.mode` has been removed.
Previously, it was a mix of a setup option (`micro_battles`), a constant
determined by the map (`replay`) and a way to tune end-of-battle detection
(`only_consider_units`). `micro_battles` and `only_consider_units` are now
considered global options, much like `window_size`. `only_consider_units` can be
changed during the game by calling `state:setOnlyConsiderUnits`. `replay` is
simply another read-only field in the state. Note that this will require changes
in existing client code; the examples have been updated accordingly.