Install headers and C++ library, put replayer and client into a single module
This re-organizes the build system and source layout so that we can install a
pure-C++ TorchCraft library with luarocks. I decide to stick to the Lua
eco-system for now to keep things simple; if needed, we could still add support
for a "standard" libtorchcraft installation later.

The source code is now organized as follows:
 - include/ contains header files that will be installed and are usable with
   non-Lua clients (i.e. using libtorchcraft only). This should probably also
   include others like gamestore.h, but there are some Lua dependencies left to
   be factored out.
 - replayer/ and client/ contain source files and internal headers for
   libtorchcraft.
 - lua/ contains Lua wrappers for code in replayer/ and client/.

There is now a single C++ Lua module containing both client and replayer
functionality, dubbed 'tc_lib'. For convenience and compatibility, replayrer.lua
and client.lua are provided which simply return the approriate table from the
C++ module.