Refactored C++ and Lua namespaces
C++ side: `client` is now `torchcraft`, and `replayer` resides in
`torchcraft::replayer`

Lua side: `torchcraft.client` is gone and the corresponding definitions have
moved to the top-level `torchcraft` module (this is done in `init.lua`).
`torchcraft.replayer` is available just like before. Importing the C++ module
will now return a table with definitions designated to the top-level module as
well as a sub-table for `replayer`.