Buildability data and enhanced unit visibility info
This adds new information to the server-side handshake: a 2D array in build tile
resolution that indicates whether a building can be placed on a given tile.

Apart from that, the visibility information of units has been enhanced.
'visibile' has been changed to an integer that stores visibility bits for each
player in the game. The result of BWAPI's isDetected() function is stored in
'detected'.