Add status flags for units
There's a lot of unit status information accessible via BWAPI as booleans, and
so far we restricted the information in TorchCraft to idle, detected and
visible. However, flags like constructing can also be helpful in certain
situations such as tracking a worker's progress.

Hence, this includes a single 64-bit integer that stores various status bits.
Furthermore, the conditions for including a unit in the frame sent to the client
have been relaxed -- it's now upon the client to ignore units that are not of
interest.

Flags are 64-bit and expected to change frequently so they are included as
absolute values in frame diffs.