Zstd compression support
This adds detection of zstd to the build system (reflected via the `WITH_ZSTD`
pre-processor symbol) and drop-in replacements for std::ifstream and
std::ofstream that read and write data using Zstd. zstd::ifstream will
automatically detected whether it is reading uncompressed data and will resort
to copy-through in this case.

Replayer is aware of Zstd support, which still has to be enabled explicitly by
the user at this point. In preliminary experiments, replay data could be
compressed by a huge factor, even when containing frame diffs.