Add command status codes to server messages
This adds a `commands_status` field to the server's Frame message. It contains a
return code for each command that the client issued in its previous message, in
the same order as the original commands. A return code of zero indicates
success, and any other value signals an error. This should be mainly useful for
debugging as it is tedious to constantly check the server's error log for failed
commands. Note that for every command issued via BWAPI, there is still a small
chance that it will fail even if its return code is zero.

The exact error codes are currently internal to the server code since I assumed
it would not be that useful to expose them on the client side. This can be
changed later, of course.

Fixes #106.