Fix possible decompression error in zstd::istreambuf
It's possible that the Zstd stream will not decompress any data but still asks
for more. I observed this in one instance in which Zstd would return '1' as a
hint for the size of the next chunk of input data, but when fed with a single
byte would not decompress anything. The previous implementation would ignore
this and pass the same pointer for all three arguments of setg(). This may cause
a segmentation fault or various other problems down the road (depending on
compiler and optimization level).