BWEnv: support IPv6 on non-Windows systems
This enables connections to BWEnv with both IPv4 and IPv6. If restricted to IPv4
only, there's an issue that appears on some non-Windows systems that prevents
establishing a connectiong with a TC client (which also supports both IPv4 and
IPv67).

I'll leave the Windows setup as is since it's functional it's a little bit of a
hassle for me to test IPv6 connectivity there.