Non-blocking mode for BWEnv
In non-blocking mode, BWEnv will block the game for a specified maximum amount
of time only before continuing. This enables abiding time limits for a given
match-up without adding complicated timing logic to the TorchCraft client
program.

If the client does not send commands for the current frame back in time (e.g.
receive in BWEnv times out), the game continues and a new attempt to receive
them will take place in the very next frame and so on until we finally get a
reply. If BWEnv absolutely has to send the frame (e.g. the game or battle is
over), it will perform a blocking receive call before sending it to abide the
REQ/REP pattern.

Please note the interplay with the SET_COMBINE_FRAMES option: BWEnv can only
send frame data to the client after having a received a message with commands.
Hence, if it is unable to send new frame data it will keep combining frames and
eventually send the combined frame in the next execution of
`Controller::onFrame()` after the successful receive. Hence, SET_COMBINE_FRAMES
defines a *minimum* number of combined frames for non-blocking mode.

The default behavior is blocking mode (i.e. the same as before). It can be
enabled specifically via the new SET_BLOCKING command. The timeout can also be
set via a new command; by default it's 50ms.