Add ability to set maximum number of combined frames
The `COMBINE_FRAMES` option takes either one or two arguments now. With two
argument, it allows setting a maximum number of combined frames. This is useful
in non-blocking mode, in which `COMBINE_FRAMES` previously acted as a *minimum*
number of combined frames. If a maximum is set, BWEnv will block and wait for a
client requests once the maximum number of combined frames has been reached.

The intended use-case is to speed up processing of replays. It's usually
desirable to have a constant number of combined frames per server reply (i.e.
setting `COMBINE_FRAMES` and enabling `BLOCKING`). However, waiting for a client
request does, in many cases, not make much sense because the client might not
issue any commands. With setting a minmum and maximum combination value  and
disabling `BLOCKING` (and setting a very low maximum frame duration), the server
will only block once a given number of new frames are ready and otherwise just
briefly check for new client requests.

I did a small benchmark with the `examples/py/dump_rep.py` and measured the
(wall clock) time on macOS that it takes to receive 5000 frames from
"maps/replays/bwrep_gwhxe.rep": 20.7s for the previous version, and 16.9s when
setting the options mentioned above (they are added to the dump_replay scripts in
the following commit). The difference should be more pronounced if client
programs are doing some actual working during replay processing.