Enable sharing of ZMQ context between active connections
For training runs, we can easily have >80 active TorchCraft connections per
process (e.g. >40 self-play games running in parallel). Constantly constructing
and destructing a ZeroMQ context for each connection is both a waste of time and
resources, since each context will result in a dedicated I/O thread. This change
enables context sharing across across *active* connections -- once the last
connection is closed, the shared context is destroyed and re-created on the
next connection attempt.

I did some basic benchmarking, and for a relatively simple self-play game I saw
500 MBit going over the wire in total (send/receive for both players) within 30
seconds. Assuming twice the size for a more complex game, we'd hit 1 Gbps with
30 parallel games. The ZeroMQ FAQ (http://zeromq.org/area:faq#toc7) suggest to
use one thread for every Gpbs so I went with two threads to be safe. In reality,
we're often performing other operations while playing, which slows down the game
and reduces the required throughput. If the game gets more complex, player steps
will be slower as well.

ZeroMQ contexts are not designed to be created and destroyed frequently anyway;
they should be retained for a longer time. On the other hand, for a library, I
still believe that it's good design to not allocate a lot of global resources
which are supposed to be cleaned up at program exit. Hence, the solution here
seems like a good compromise.

This also includes setting the "linger" period for connection sockets. This is
infinite by default, but now that the context can outlive the connection we
don't want pending messages to linger in the context for too long.