Prevent concurrent connections from multiple clients
While it's possible to reconnect to a server and continue playing, controlling a
player with multiple clients at once is a strong indication that something is
wrong. It also results in clients receiving wrong (frame diff) data.

The ZMQ server will now keep track of the last client that performed a handshake
and will only accept commands from this client. Commands from any other client
will result in error messages being sent back. In blocking mode, the game will
not advance in this case; in non-blocking mode, the game will advance by one
frame but no commands will be executed.

With some refactoring, it would be possible to make this work correctly in
non-blocking mode as well. However, concurrent connections are considered an
error now (likely indicating another error in the client setup) and hence I
think that 100% correct operation will not be required.